Updates, clean-ups, and fixes (#302)
* General clean-up

- factored out common functions from cows and mfhelper
- add AreaWatcher.js for droppers/mules
- changed Overrides_1 to Overrides to follow format used in entrys
- fixed loader.js bug, if a script failed when there was a reconfiguration applied, the original config was never copied back.
- add script duration printing to loader
- fixed Pickit.js error if config wasn't loaded, i.e autmule, mulelogger
- re-ordered town chores, better scroll handling, fix crash during buyPotions
- factored out itemEvent from PickThread and default.dbj to Pickit.js

* Include Worldstone in mfhelper leader check

- clean up, missed deleting an unused i variable
- fix typo in sdk and Pather for ForgottenSands

* Change getNearestMonster to use obj settings

- use object settings instead of multiple params as its easier to understand, add skipGid

* Better potion managment in town

- handle clearing potions better, prevent crashes from trying to move a potion where we couldn't or having something on our cursor while moving potions around
- make decidedSkill for sorceress classattack. Might do it for others but current just sorc so she doesn't stand there doing nothing while merc stomping

* Script clean-up

- mostly just minimizing code
- factored out Ethic's autoLeaderDetect to be a function in Misc

* A lot really

- A lot of general clean-up
- More console logging information, still need more so we can actually track what bot is doing and why
- removed depreciated Attack.checkMonster in favor of attackable prototype
- removed depreciated Attack.weaponSwitch in favor of new switchWeapons prototype
- added new Precast.precastables object, to track what skills we have rather than checking it them every time doPrecast is called, changed autobuild to re-initialize the data if we level. Created Skill.init function
- changed handing of ColdArmor during precast
- factored out whirlwind from barb/sin classattack
- added checkNearCorpses and checkCorpse to Attack. Currently just being used by paladin to ensure there are corpses around when he's trying to use redemption
- finally fixed Common.Diablo.openSeal to work with de-seis's bugged seal
- removed Pickit.useTelekinesis, pointless to have Pickit have it and Skill have it, added Skill.haveTk -> gets initialized on game join and re-checked everytime Skill.init is called
- added MinColumn check to ensure user didn't incorrectly input MinColumn to be more than their current belt can handle
- added Pickit.result obj for clarity of code
- started process of changing pather settings to use objects so we can better control bot actions
- initialize wp data on game join
- added nonTownWpAreas
- handle cleaning up potions better
- added clickUnitAndWait to Util.js, so we can actually know if the unit we clicked on was interacted with

* Update Pickit.js

- remove line that got added back after updating from master

* Stability updates

- change some of the default behavior for precasting/potting/picking to use packets
- added more delays and removed some of the me.ping checks as they seem to cause issues, (less crashes over a 24hr period)
- handle getUnit bug when attempting to go to town
- re-wrote Precast.precastSkill to be its own stand-alone casting function with optional weapon switching rather than old method with wasteful checks. todo still is track precastables durations
- fix check in pickItem, old method didn't account for failing to moveToUnit if we had teleport
- use packet first by default then clickItem if it fails and we aren't using fastPick
- change findSpotOnDistance to use Pather.checkSpot rather than Attack.validSpot, wanted to check if the spot is on the floor. Was getting issues at diablo
- check if node we are moving to is a valid spot if we fail to moveTo or Near it 3 times and try to get nearestWalkable node if it isn't
- add delay in usePortal if portal was a red portal
- wrote Skill.getDuration to calculate duration of precast skills
- wrote Packet.givePotToMerc, can be used to stack thawing/antidote potions on the merc from our inventory or using of hp/rv buffer potions on them. More testing is needed to see if that is helpful
- added default zeal/vengeance attack sequence to paladin classattacks, don't need to check skill range ect and we need an additional collision check to make sure we aren't blocked by walls. Still need work for cases like abddaon where monster is on other side of blockade
- remove AnniStarter.dbj check from AutoMule and D2BotShopDrop from default.dbj
- fix missing variable declarations in D2BotMule.dbj

* Update Town.js

- fix typo in buyPots

* Add monster objects exception to attackable prototype

- let the bot attack monster objects, barricades and doors. They were returning as an invalid attack location causing them to be skipped during classattack and are included in neverCount so were returning false for attackable
- fixed a5 getPath crash during usePortal
- wrapped extra options inside of Eldritch.js in try/catch
- added anti-idle to Gamble.js

* Add character name in updateStatus

- easier to tell which profile belongs to which character

* remove duplicated functions from gameaction/mulelogger

- keep Mule logging functions in MuleLogger, call it as needed from GameAction

* Update Town.js

- last update for now I think, minor changes mostly just some clean up
- changed order of checks for getTpTool to save some time

+++ d2bs/kolbot/libs/bots/Diablo.js
@@ -58,15 +58,11 @@ function Diablo() {
 		print("Seis layout " + Common.Diablo.seisLayout);
 		this.followPath(Common.Diablo.seisLayout === 1 ? this.starToSeisA : this.starToSeisB);
 
-		if (!Common.Diablo.openSeal(394)) {
-			throw new Error("Failed to open de Seis seal.");
-		}
+		if (!Common.Diablo.openSeal(394)) throw new Error("Failed to open de Seis seal.");
 
 		Common.Diablo.seisLayout === 1 ? Pather.moveTo(7798, 5194) : Pather.moveTo(7796, 5155);
 
-		if (!this.getBoss(getLocaleString(2852))) {
-			throw new Error("Failed to kill de Seis");
-		}
+		if (!this.getBoss(getLocaleString(2852))) throw new Error("Failed to kill de Seis");
 
 		return true;
 	};
@@ -76,9 +72,7 @@ function Diablo() {
 		print("Inf layout " + Common.Diablo.infLayout);
 		this.followPath(Common.Diablo.infLayout === 1 ? this.starToInfA : this.starToInfB);
 
-		if (!Common.Diablo.openSeal(392)) {
-			throw new Error("Failed to open Infector seals.");
-		}
+		if (!Common.Diablo.openSeal(392)) throw new Error("Failed to open Infector seals.");
 
 		if (Common.Diablo.infLayout === 1) {
 			if (me.sorceress || me.assassin) {
@@ -91,13 +85,8 @@ function Diablo() {
 			Pather.moveTo(7928, 5295);
 		}
 
-		if (!this.getBoss(getLocaleString(2853))) {
-			throw new Error("Failed to kill Infector");
-		}
-
-		if (!Common.Diablo.openSeal(393)) {
-			throw new Error("Failed to open Infector seals.");
-		}
+		if (!this.getBoss(getLocaleString(2853))) throw new Error("Failed to kill Infector");
+		if (!Common.Diablo.openSeal(393)) throw new Error("Failed to open Infector seals.");
 
 		return true;
 	};
@@ -139,7 +128,7 @@ function Diablo() {
 
 		if (monster) {
 			do {
-				if (Attack.checkMonster(monster)) {
+				if (monster.attackable) {
 					for (let i = 0; i < this.cleared.length; i += 1) {
 						if (getDistance(monster, this.cleared[i][0], this.cleared[i][1]) < 30 && Attack.validSpot(monster.x, monster.y)) {
 							Pather.moveToUnit(monster);
@@ -152,9 +141,7 @@ function Diablo() {
 			} while (monster.getNext());
 		}
 
-		if (getDistance(me, oldPos.x, oldPos.y) > 5) {
-			Pather.moveTo(oldPos.x, oldPos.y);
-		}
+		getDistance(me, oldPos.x, oldPos.y) > 5 && Pather.moveTo(oldPos.x, oldPos.y);
 
 		return true;
 	};
@@ -180,9 +167,7 @@ function Diablo() {
 			} while (monster.getNext());
 		}
 
-		if (getDistance(me, oldPos.x, oldPos.y) > 5) {
-			Pather.moveTo(oldPos.x, oldPos.y);
-		}
+		getDistance(me, oldPos.x, oldPos.y) > 5 && Pather.moveTo(oldPos.x, oldPos.y);
 
 		return true;
 	};
@@ -193,9 +178,7 @@ function Diablo() {
 	Precast.doPrecast(true);
 	me.area !== sdk.areas.RiverofFlame && Pather.useWaypoint(sdk.areas.RiverofFlame);
 
-	if (!Pather.moveTo(7790, 5544)) {
-		throw new Error("Failed to move to Chaos Sanctuary");
-	}
+	if (!Pather.moveToExit(sdk.areas.ChaosSanctuary, true) && !Pather.moveTo(7790, 5544)) throw new Error("Failed to move to Chaos Sanctuary");
 
 	Common.Diablo.initLayout();
